Use floating point 0 - 1 when dealing with rgb colors. All code updated except when dealing directly with vertices (i.e. in the bias lighting code).
Found and fixed an issue with the light grid which resulted in evaluated points returning {0, 0, 0} if the evaluation happens after a change but before an update. In this instance we copy the old value prior to being marked for update (not entirely accurate but given that the next update is only a single tic away; its OK).

Part #1

+++ doomsday/engine/portable/include/p_maptypes.h
@@ -91,8 +91,8 @@ typedef struct surface_s {
     float               oldoffx;
     float               offy;          // Texture y offset
     float               oldoffy;
-    byte                rgba[4];       // Surface color tint
-    byte                oldrgba[4];
+    float               rgba[4];       // Surface color tint
+    float               oldrgba[4];
     struct translation_s* xlat;
 } surface_t;
 
@@ -112,7 +112,7 @@ typedef struct plane_s {
     float               oldheight[2];
     surface_t           surface;
     float               glow;          // Glow amount
-    byte                glowrgb[3];    // Glow color
+    float               glowrgb[3];    // Glow color
     float               target;        // Target height
     float               speed;         // Move speed
     degenmobj_t         soundorg;      // Sound origin for plane
@@ -174,8 +174,8 @@ typedef struct sector_s {
     runtime_mapdata_header_t header;
     float               lightlevel;
     float               oldlightlevel;
-    byte                rgb[3];
-    byte                oldrgb[3];
+    float               rgb[3];
+    float               oldrgb[3];
     int                 validcount;    // if == validcount, already checked.
     struct mobj_s*      thinglist;     // List of mobjs in the sector.
     unsigned int        linecount;
